Infrared obstacle avoidance sensors
This Infrared Obstacle Avoidance Sensor returns a signal when it detects an object in range. The range of the sensor is around 2-40 cm is distance. It operates at 3.5 to 5 volts at around 20 milliamps. The Obstacle Avoidance Sensors usually come in two types - with 3 and 4 pins. The 3 pin version does not have the ability to be enabled/disabled. The 4 pin version has optional Enable pin. Infrared obstacle avoidance sensor is designed to detect obstacles or the difference in reflective services.