Rotary encoder module
Rotation of the rotary encoder can be counted if it is rotated to the positive direction and the opposite direction outputs pulse frequency, unlike rotary potentiometer counter, which the rotation counts are not limited. With the buttons on the rotary encoder can be reset to its initial state, that starts counting from 0.